+++ src/Filament/Resources/Transactions/Schemas/TransactionInfolist.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace OfficeGuy\LaravelSumitGateway\Filament\Resources\Transactions\Schemas;
+
+use Filament\Schemas\Schema;
+
+class TransactionInfolist
+{
+    public static function configure(Schema $schema): Schema
+    {
+        return $schema
+            ->components([
+                //
+            ]);
+    }
+}